CarSaleContract
The test I wrote with Javascript inside is to check that an Ethereum contract called CarSale is working correctly. First, the contract is deployed and then we run three tests. The tests check that the buyer and sale price information is updated correctly when selling a car, that only the seller can sell a car, and that the sale price must be the minimum car price. These tests were written to check that the contract correctly handles car for sale transactions.